+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Unreleased
+## [1.0.0-rc1] - 2021-05-03
 ### Added
 - `K8s.Resource.NamedList.access!/1` raises if item is missing
 - K8s.Operation.put_label_selector/2